Time: 6 ms (86.75%), Space: 10.5 MB (77.91%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -59,7 +59,7 @@ public:
             }
         }
 
-        return dp[amount];
+        return (int) dp[amount];
     }
 
 };
